requirements.txt
# vertex / kfp core
google-cloud-aiplatform==1.94.0     # Vertex AI
google-cloud-pipeline-components==2.20.0
kfp==2.13.0

# plumbing
google-cloud-storage==2.16.0
protobuf<4.26

# add your model libs underneath
numpy
pandas
scikit-learn

